requirements.txt
pytest==4.6.9
pytest-sugar==0.9.2
pytest-xdist==1.31.0
factory-boy==2.12.0